Update README.md
We saw that Sales column contains 172817 rows with 0 sale. So, we created a new data frame in which we removed 0 sales rows and tried to train our model. We used various algorithms and got accuracy score around 74%.

We were also curious about the total dataset (including Sales = 0 rows). So, we trained another model using various algorithms and we got accuracy near about 98% which is far better than previous model.

So, we came to conclusion that removing sales=0 rows actually remove lot of information from dataset as it has 172817 rows which is quite large and therefore, we decided not to remove those values. We got our best rmpse score from Random Forest model, Gradient boosting technique like adaboost, Xgboost, we tried taking an optimum parameter so that our model doesn't overfit.